refactoring: add SiteChecker class
