+++ index.ts
@@ -1,3 +1,8 @@
-import { siteChecker, SiteCheckerOptions } from './lib/site-checker';
+import { SiteChecker, SiteInfo, SiteCheckerOptions } from './lib/SiteChecker';
 
-export { siteChecker, SiteCheckerOptions };
+export async function siteChecker(list: SiteInfo[], opts: SiteCheckerOptions): Promise<SiteInfo[]> {
+  const sc = new SiteChecker(list, opts);
+  return sc.run();
+}
+
+export { SiteChecker, SiteInfo, SiteCheckerOptions };
